Refactoring: Create new ValidationRequirement
